Refund

+++ egov/egov-egf/src/main/java/org/egov/model/report/DedicatedExpenseViewData.java
@@ -1,97 +1,95 @@
 package org.egov.model.report;
 
+import java.math.BigDecimal;
+
 public class DedicatedExpenseViewData {
 	private String name;
 	
-	private Double glsum1;
-	private Double glsum2;
-	private Double glsum3;
-	private Double glsum4;
-	private Double glsum5;
-	private Double glsum6;
-	private Double glsum7;
-	private Double glsum8;
-	private Double glsum9;
-	private Double glsum10;
-	private Double sum;
-	/* private Double colsum1; */
-	
-	/*
-	 * public Double getColsum1() { return colsum1; } public void setColsum1(Double
-	 * colsum1) { this.colsum1 = colsum1; }
-	 */
+	private BigDecimal glsum1;
+	private BigDecimal glsum2;
+	private BigDecimal glsum3;
+	private BigDecimal glsum4;
+	private BigDecimal glsum5;
+	private BigDecimal glsum6;
+	private BigDecimal glsum7;
+	private BigDecimal glsum8;
+	private BigDecimal glsum9;
+	private BigDecimal glsum10;
+	private BigDecimal sum;
 	public String getName() {
 		return name;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
-	public Double getGlsum1() {
+	public BigDecimal getGlsum1() {
 		return glsum1;
 	}
-	public void setGlsum1(Double glsum1) {
+	public void setGlsum1(BigDecimal glsum1) {
 		this.glsum1 = glsum1;
 	}
-	public Double getGlsum2() {
+	public BigDecimal getGlsum2() {
 		return glsum2;
 	}
-	public void setGlsum2(Double glsum2) {
+	public void setGlsum2(BigDecimal glsum2) {
 		this.glsum2 = glsum2;
 	}
-	public Double getGlsum3() {
+	public BigDecimal getGlsum3() {
 		return glsum3;
 	}
-	public void setGlsum3(Double glsum3) {
+	public void setGlsum3(BigDecimal glsum3) {
 		this.glsum3 = glsum3;
 	}
-	public Double getGlsum4() {
+	public BigDecimal getGlsum4() {
 		return glsum4;
 	}
-	public void setGlsum4(Double glsum4) {
+	public void setGlsum4(BigDecimal glsum4) {
 		this.glsum4 = glsum4;
 	}
-	public Double getGlsum5() {
+	public BigDecimal getGlsum5() {
 		return glsum5;
 	}
-	public void setGlsum5(Double glsum5) {
+	public void setGlsum5(BigDecimal glsum5) {
 		this.glsum5 = glsum5;
 	}
-	public Double getGlsum6() {
+	public BigDecimal getGlsum6() {
 		return glsum6;
 	}
-	public void setGlsum6(Double glsum6) {
+	public void setGlsum6(BigDecimal glsum6) {
 		this.glsum6 = glsum6;
 	}
-	public Double getGlsum7() {
+	public BigDecimal getGlsum7() {
 		return glsum7;
 	}
-	public void setGlsum7(Double glsum7) {
+	public void setGlsum7(BigDecimal glsum7) {
 		this.glsum7 = glsum7;
 	}
-	public Double getGlsum8() {
+	public BigDecimal getGlsum8() {
 		return glsum8;
 	}
-	public void setGlsum8(Double glsum8) {
+	public void setGlsum8(BigDecimal glsum8) {
 		this.glsum8 = glsum8;
 	}
-	public Double getGlsum9() {
+	public BigDecimal getGlsum9() {
 		return glsum9;
 	}
-	public void setGlsum9(Double glsum9) {
+	public void setGlsum9(BigDecimal glsum9) {
 		this.glsum9 = glsum9;
 	}
-	public Double getGlsum10() {
+	public BigDecimal getGlsum10() {
 		return glsum10;
 	}
-	public void setGlsum10(Double glsum10) {
+	public void setGlsum10(BigDecimal glsum10) {
 		this.glsum10 = glsum10;
 	}
-	public Double getSum() {
+	public BigDecimal getSum() {
 		return sum;
 	}
-	public void setSum(Double sum) {
+	public void setSum(BigDecimal sum) {
 		this.sum = sum;
 	}
+	
+
